+++ CHANGELOG.md
@@ -4,6 +4,9 @@ Changelog
 2.0.0
 -----
 
+* **2015-05-29**: [BC BREAK] Deleted the `sonata_admin` bundle set
+* **2015-05-29**: [BC BREAK] Deleted the `LoadBaseData` data fixtures
+* **2015-05-29**: [BC BREAK] Deleted the `Content` document
 * **2015-05-29**: [BC BREAK] Moved `BaseTestCase#assertResponseOk()` to `Assert#responseOk()`
 
 1.2.0-RC2